+++ internal/backend/local/test.go
@@ -13,17 +13,16 @@ import (
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/zclconf/go-cty/cty"
-	"github.com/zclconf/go-cty/cty/convert"
 	"golang.org/x/exp/slices"
 
 	"github.com/hashicorp/terraform/internal/addrs"
 	"github.com/hashicorp/terraform/internal/backend"
 	"github.com/hashicorp/terraform/internal/command/views"
 	"github.com/hashicorp/terraform/internal/configs"
-	"github.com/hashicorp/terraform/internal/lang"
-	"github.com/hashicorp/terraform/internal/lang/marks"
 	"github.com/hashicorp/terraform/internal/logging"
 	"github.com/hashicorp/terraform/internal/moduletest"
+	configtest "github.com/hashicorp/terraform/internal/moduletest/config"
+	hcltest "github.com/hashicorp/terraform/internal/moduletest/hcl"
 	"github.com/hashicorp/terraform/internal/plans"
 	"github.com/hashicorp/terraform/internal/states"
 	"github.com/hashicorp/terraform/internal/terraform"
@@ -98,6 +97,14 @@ func (runner *TestSuiteRunner) Test() (moduletest.Status, tfdiags.Diagnostics) {
 
 		file := suite.Files[name]
 
+		priorStates := make(map[string]*terraform.TestContext)
+		for _, run := range file.Runs {
+			// Pre-initialise the prior states, so we can easily tell between
+			// a run block that doesn't exist and a run block that hasn't been
+			// executed yet.
+			priorStates[run.Name] = nil
+		}
+
 		fileRunner := &TestFileRunner{
 			Suite: runner,
 			RelevantStates: map[string]*TestFileState{
@@ -106,7 +113,7 @@ func (runner *TestSuiteRunner) Test() (moduletest.Status, tfdiags.Diagnostics) {
 					State: states.NewState(),
 				},
 			},
-			PriorStates: make(map[string]*terraform.TestContext),
+			PriorStates: priorStates,
 		}
 
 		runner.View.File(file, moduletest.Starting)
@@ -216,6 +223,8 @@ type TestFileRunner struct {
 	// validate the test assertions, and used when calculating values for
 	// variables within run blocks.
 	PriorStates map[string]*terraform.TestContext
+
+	globalVariables map[string]backend.UnparsedVariableValue
 }
 
 // TestFileState is a helper struct that just maps a run block to the state that
@@ -227,6 +236,9 @@ type TestFileState struct {
 
 func (runner *TestFileRunner) Test(file *moduletest.File) {
 	log.Printf("[TRACE] TestFileRunner: executing test file %s", file.Name)
+
+	// First thing, initialise the global variables for the file
+	runner.initVariables(file)
 
 	// We'll execute the tests in the file. First, mark the overall status as
 	// being skipped. This will ensure that if we've cancelled and the files not
@@ -337,7 +349,7 @@ func (runner *TestFileRunner) run(run *moduletest.Run, file *moduletest.File, st
 		return state, false
 	}
 
-	resetConfig, configDiags := config.TransformForTest(run.Config, file.Config)
+	resetConfig, configDiags := configtest.TransformConfigForTest(config, run, file, runner.globalVariables)
 	defer resetConfig()
 
 	run.Diagnostics = run.Diagnostics.Append(configDiags)
@@ -360,7 +372,14 @@ func (runner *TestFileRunner) run(run *moduletest.Run, file *moduletest.File, st
 		return state, false
 	}
 
-	planCtx, plan, planDiags := runner.plan(config, state, run, file, references, start)
+	variables, variableDiags := runner.GetVariables(config, run, file, references)
+	run.Diagnostics = run.Diagnostics.Append(variableDiags)
+	if variableDiags.HasErrors() {
+		run.Status = moduletest.Error
+		return state, false
+	}
+
+	planCtx, plan, planDiags := runner.plan(config, state, run, file, runner.FilterVariablesToConfig(config, variables), references, start)
 	if run.Config.Command == configs.PlanTestCommand {
 		// Then we want to assess our conditions and diagnostics differently.
 		planDiags = run.ValidateExpectedFailures(planDiags)
@@ -370,14 +389,8 @@ func (runner *TestFileRunner) run(run *moduletest.Run, file *moduletest.File, st
 			return state, false
 		}
 
-		variables, resetVariables, variableDiags := runner.prepareInputVariablesForAssertions(config, run, file, references)
+		resetVariables := runner.AddVariablesToConfig(config, variables)
 		defer resetVariables()
-
-		run.Diagnostics = run.Diagnostics.Append(variableDiags)
-		if variableDiags.HasErrors() {
-			run.Status = moduletest.Error
-			return state, false
-		}
 
 		if runner.Suite.Verbose {
 			schemas, diags := planCtx.Schemas(config, plan.PlannedState)
@@ -456,10 +469,8 @@ func (runner *TestFileRunner) run(run *moduletest.Run, file *moduletest.File, st
 		return updated, true
 	}
 
-	variables, resetVariables, variableDiags := runner.prepareInputVariablesForAssertions(config, run, file, references)
-	if resetVariables != nil {
-		defer resetVariables()
-	}
+	resetVariables := runner.AddVariablesToConfig(config, variables)
+	defer resetVariables()
 
 	run.Diagnostics = run.Diagnostics.Append(variableDiags)
 	if variableDiags.HasErrors() {
@@ -552,7 +563,7 @@ func (runner *TestFileRunner) destroy(config *configs.Config, state *states.Stat
 
 	var diags tfdiags.Diagnostics
 
-	variables, variableDiags := runner.buildInputVariablesForTest(run, file, config)
+	variables, variableDiags := runner.GetVariables(config, run, file, nil)
 	diags = diags.Append(variableDiags)
 
 	if diags.HasErrors() {
@@ -561,7 +572,7 @@ func (runner *TestFileRunner) destroy(config *configs.Config, state *states.Stat
 
 	planOpts := &terraform.PlanOpts{
 		Mode:         plans.DestroyMode,
-		SetVariables: variables,
+		SetVariables: runner.FilterVariablesToConfig(config, variables),
 	}
 
 	tfCtx, ctxDiags := terraform.NewContext(runner.Suite.Opts)
@@ -603,7 +614,7 @@ func (runner *TestFileRunner) destroy(config *configs.Config, state *states.Stat
 	return updated, diags
 }
 
-func (runner *TestFileRunner) plan(config *configs.Config, state *states.State, run *moduletest.Run, file *moduletest.File, references []*addrs.Reference, start int64) (*terraform.Context, *plans.Plan, tfdiags.Diagnostics) {
+func (runner *TestFileRunner) plan(config *configs.Config, state *states.State, run *moduletest.Run, file *moduletest.File, variables terraform.InputValues, references []*addrs.Reference, start int64) (*terraform.Context, *plans.Plan, tfdiags.Diagnostics) {
 	log.Printf("[TRACE] TestFileRunner: called plan for %s/%s", file.Name, run.Name)
 
 	var diags tfdiags.Diagnostics
@@ -613,9 +624,6 @@ func (runner *TestFileRunner) plan(config *configs.Config, state *states.State, 
 
 	replaces, replaceDiags := run.GetReplaces()
 	diags = diags.Append(replaceDiags)
-
-	variables, variableDiags := runner.buildInputVariablesForTest(run, file, config)
-	diags = diags.Append(variableDiags)
 
 	if diags.HasErrors() {
 		return nil, nil, diags
@@ -837,7 +845,7 @@ func (runner *TestFileRunner) cleanup(file *moduletest.File) {
 			diags = diags.Append(tfdiags.Sourceless(tfdiags.Error, "Inconsistent state", fmt.Sprintf("Found inconsistent state while cleaning up %s. This is a bug in Terraform - please report it", file.Name)))
 		}
 	} else {
-		reset, configDiags := runner.Suite.Config.TransformForTest(main.Run.Config, file.Config)
+		reset, configDiags := configtest.TransformConfigForTest(runner.Suite.Config, main.Run, file, runner.globalVariables)
 		diags = diags.Append(configDiags)
 
 		if !configDiags.HasErrors() {
@@ -912,7 +920,7 @@ func (runner *TestFileRunner) cleanup(file *moduletest.File) {
 
 		var diags tfdiags.Diagnostics
 
-		reset, configDiags := state.Run.Config.ConfigUnderTest.TransformForTest(state.Run.Config, file.Config)
+		reset, configDiags := configtest.TransformConfigForTest(state.Run.Config.ConfigUnderTest, state.Run, file, runner.globalVariables)
 		diags = diags.Append(configDiags)
 
 		updated := state.State
@@ -933,136 +941,149 @@ func (runner *TestFileRunner) cleanup(file *moduletest.File) {
 	}
 }
 
-// buildInputVariablesForTest creates a terraform.InputValues mapping for
-// variable values that are relevant to the config being tested.
+// GetVariables builds the terraform.InputValues required for the provided run
+// block. It pulls the relevant variables (ie. the variables needed for the
+// run block) from the total pool of all available variables, and converts them
+// into input values.
 //
-// Crucially, it differs from prepareInputVariablesForAssertions in that it only
-// includes variables that are reference by the config and not everything that
-// is defined within the test run block and test file.
-func (runner *TestFileRunner) buildInputVariablesForTest(run *moduletest.Run, file *moduletest.File, config *configs.Config) (terraform.InputValues, tfdiags.Diagnostics) {
+// As a run block can reference variables defined within the file and are not
+// actually defined within the configuration, this function actually returns
+// more variables than are required by the config. FilterVariablesToConfig
+// should be called before trying to use these variables within a Terraform
+// plan, apply, or destroy operation.
+func (runner *TestFileRunner) GetVariables(config *configs.Config, run *moduletest.Run, file *moduletest.File, references []*addrs.Reference) (terraform.InputValues, tfdiags.Diagnostics) {
 	var diags tfdiags.Diagnostics
 
-	// configVariables keeps track of the variables that will actually be given
-	// to the terraform graph to provide values to the configuration.
-	configVariables := make(terraform.InputValues)
+	// relevantVariables contains the variables that are of interest to this
+	// run block. We can have variables defined at the global level and at the
+	// file level that this run block doesn't need so we're going to make a
+	// quick list of the variables that are actually relevant.
+	relevantVariables := make(map[string]bool)
 
-	// ctxVariables contains all the possible variables we have definitions for
-	// and is used to build the context that is used to evaluate variables.
-	ctxVariables := make(terraform.InputValues)
-
-	// First, we process all the global variables.
-	for name, value := range runner.Suite.GlobalVariables {
-		var variableDiags tfdiags.Diagnostics
-		if variable, exists := config.Module.Variables[name]; exists {
-			ctxVariables[name], variableDiags = value.ParseVariableValue(variable.ParsingMode)
-			configVariables[name] = ctxVariables[name]
-		} else {
-			// Since we don't have the config here to parse the variable value
-			// we just blanket parse it as an HCL expression. We don't include
-			// this in the configVariables, as we only want variables that are
-			// defined in the context.
-			ctxVariables[name], variableDiags = value.ParseVariableValue(configs.VariableParseHCL)
-		}
-		diags = diags.Append(variableDiags)
-	}
-
-	// Second, we process the variables defined at the file level
-	//
-	// We're happy for anything here to override any values from the global
-	// variables.
-	if file != nil {
-		for name, expr := range file.Config.Variables {
-
-			value := unparsedTestVariableValue{
-				expr: expr,
-			}
-
-			var variableDiags tfdiags.Diagnostics
-			if variable, exists := config.Module.Variables[name]; exists {
-				ctxVariables[name], variableDiags = value.ParseVariableValue(variable.ParsingMode)
-				configVariables[name] = ctxVariables[name]
-			} else {
-				// As above, we don't have this defined in the config so we
-				// parse it as an expression and don't include it in
-				// configVariables.
-				ctxVariables[name], variableDiags = value.ParseVariableValue(configs.VariableParseHCL)
-			}
-			diags = diags.Append(variableDiags)
+	// First, we'll check to see which variables the run block assertions
+	// reference.
+	for _, reference := range references {
+		if addr, ok := reference.Subject.(addrs.InputVariable); ok {
+			relevantVariables[addr.Name] = true
 		}
 	}
 
-	// Thirdly, we process the variables defined at the run level and pull out
-	// any that are relevant to the config under test.
-	//
-	// We're happy for anything here to override any values from the global or
-	// file level variables
-	if run != nil {
-		skipVars := false
-
-		ctx, ctxDiags := runner.ctx(run, file, ctxVariables)
-		diags = diags.Append(ctxDiags)
-		if ctxDiags.HasErrors() {
-			// We still want to validate all the right variables are being
-			// declared. So we don't return early, but we note that we shouldn't
-			// eval vars from this block.
-			skipVars = true
-		}
-
-		for name, expr := range run.Config.Variables {
-			variable, exists := config.Module.Variables[name]
-			if !exists {
-				// At this point we are going to add a warning if a variable
-				// is defined within a run block and not referenced by the
-				// configuration under test.
-				diags = diags.Append(&hcl.Diagnostic{
-					Severity: hcl.DiagWarning,
-					Summary:  "Value for undeclared variable",
-					Detail:   fmt.Sprintf("The module under test does not declare a variable named %q, but it is declared in run block %q.", name, run.Name),
-					Subject:  expr.Range().Ptr(),
-				})
-
-				continue
-			}
-
-			if skipVars {
-				// Then we don't have a valid evaluation context, so we won't
-				// actually process these variables. We'll put in a dummy value
-				// knowing that we have errors in the diags so these won't be
-				// processed.
-				//
-				// We still want to track this variable has a value, even if we
-				// don't know what it is, because we have some validations later
-				// that we don't want to trigger because this variable is
-				// missing.
-
-				configVariables[name] = &terraform.InputValue{
-					Value:       cty.NilVal,
-					SourceType:  terraform.ValueFromConfig,
-					SourceRange: tfdiags.SourceRangeFromHCL(variable.DeclRange),
+	// Second, we'll check to see which variables the run block variables
+	// themselves reference. We might be processing variables just for the file
+	// so the run block itself could be nil.
+	for _, expr := range run.Config.Variables {
+		for _, variable := range expr.Variables() {
+			reference, referenceDiags := addrs.ParseRefFromTestingScope(variable)
+			diags = diags.Append(referenceDiags)
+			if reference != nil {
+				if addr, ok := reference.Subject.(addrs.InputVariable); ok {
+					relevantVariables[addr.Name] = true
 				}
-
-				continue
 			}
-
-			value := unparsedTestVariableValue{
-				expr: expr,
-				ctx:  ctx,
-			}
-
-			var variableDiags tfdiags.Diagnostics
-			configVariables[name], variableDiags = value.ParseVariableValue(variable.ParsingMode)
-			diags = diags.Append(variableDiags)
 		}
-
 	}
 
-	// Finally, we'll do something about any variables defined in the
-	// configuration that we haven't given values for.
+	// Finally, we'll check to see which variables are actually defined within
+	// the configuration.
+	for name := range config.Module.Variables {
+		relevantVariables[name] = true
+	}
+
+	// Now we know which variables are actually needed by this run block.
+
+	// We're going to run over all the sets of variables we have access to:
+	//   - Global variables, from the CLI / env vars / .tfvars files.
+	//   - File variables, defined within the `variables` block in the file.
+	//   - Run variables, defined within the `variables` block in this run.
+	//   - ConfigVariables variables, defined directly within the config.
+	values := make(terraform.InputValues)
+
+	// First, let's look at the global variables.
+	for name, value := range runner.globalVariables {
+		if !relevantVariables[name] {
+			// Then this run block doesn't need this value.
+			continue
+		}
+
+		// By default, we parse global variables as HCL inputs.
+		parsingMode := configs.VariableParseHCL
+
+		cfg, exists := config.Module.Variables[name]
+		if exists {
+			// Unless we have some configuration that can actually tell us
+			// what parsing mode to use.
+			parsingMode = cfg.ParsingMode
+		}
+
+		var valueDiags tfdiags.Diagnostics
+		values[name], valueDiags = value.ParseVariableValue(parsingMode)
+		diags = diags.Append(valueDiags)
+	}
+
+	// Second, we'll check the run level variables.
+
+	// This is a bit more complicated, as the run level variables can reference
+	// previously defined variables.
+
+	// Preload the available expressions, we're going to validate them when we
+	// build the context.
+	var exprs []hcl.Expression
+	for _, expr := range run.Config.Variables {
+		exprs = append(exprs, expr)
+	}
+
+	// Preformat the variables we've processed already - these will be made
+	// available to the eval context.
+	variables := make(map[string]cty.Value)
+	for name, value := range values {
+		variables[name] = value.Value
+	}
+
+	ctx, ctxDiags := hcltest.EvalContext(exprs, variables, runner.PriorStates)
+	diags = diags.Append(ctxDiags)
+
+	var failedContext bool
+	if ctxDiags.HasErrors() {
+		// If we couldn't build the context, we won't actually process these
+		// variables. Instead, we'll fill them with an empty value but still
+		// make a note that the user did provide them.
+		failedContext = true
+	}
+
+	for name, expr := range run.Config.Variables {
+		if !relevantVariables[name] {
+			// We'll add a warning for this. Since we're right in the run block
+			// users shouldn't be defining variables that are not relevant.
+			diags = diags.Append(&hcl.Diagnostic{
+				Severity: hcl.DiagWarning,
+				Summary:  "Value for undeclared variable",
+				Detail:   fmt.Sprintf("The module under test does not declare a variable named %q, but it is declared in run block %q.", name, run.Name),
+				Subject:  expr.Range().Ptr(),
+			})
+			continue
+		}
+
+		value := cty.NilVal
+		if !failedContext {
+			var valueDiags hcl.Diagnostics
+			value, valueDiags = expr.Value(ctx)
+			diags = diags.Append(valueDiags)
+		}
+
+		values[name] = &terraform.InputValue{
+			Value:       value,
+			SourceType:  terraform.ValueFromConfig,
+			SourceRange: tfdiags.SourceRangeFromHCL(expr.Range()),
+		}
+	}
+
+	// Finally, we check the configuration again. This is where we'll discover
+	// if there's any missing variables and fill in any optional variables that
+	// don't have a value already.
 
 	for name, variable := range config.Module.Variables {
-
-		if _, exists := configVariables[name]; exists {
-			// Then we have a value for this variable already.
+		if _, exists := values[name]; exists {
+			// Then we've provided a variable for this. It's all good.
 			continue
 		}
 
@@ -1071,7 +1092,6 @@ func (runner *TestFileRunner) buildInputVariablesForTest(run *moduletest.Run, fi
 		// if they have one.
 
 		if variable.Required() {
-
 			diags = diags.Append(&hcl.Diagnostic{
 				Severity: hcl.DiagError,
 				Summary:  "No value for required variable",
@@ -1080,13 +1100,13 @@ func (runner *TestFileRunner) buildInputVariablesForTest(run *moduletest.Run, fi
 				Subject: variable.DeclRange.Ptr(),
 			})
 
-			configVariables[name] = &terraform.InputValue{
+			values[name] = &terraform.InputValue{
 				Value:       cty.DynamicVal,
 				SourceType:  terraform.ValueFromConfig,
 				SourceRange: tfdiags.SourceRangeFromHCL(variable.DeclRange),
 			}
 		} else {
-			configVariables[name] = &terraform.InputValue{
+			values[name] = &terraform.InputValue{
 				Value:       cty.NilVal,
 				SourceType:  terraform.ValueFromConfig,
 				SourceRange: tfdiags.SourceRangeFromHCL(variable.DeclRange),
@@ -1095,186 +1115,36 @@ func (runner *TestFileRunner) buildInputVariablesForTest(run *moduletest.Run, fi
 
 	}
 
-	return configVariables, diags
+	return values, diags
 }
 
-// prepareInputVariablesForAssertions creates a terraform.InputValues mapping
-// that contains all the variables defined for a given run and file, alongside
-// any unset variables that have defaults within the provided config.
+// FilterVariablesToConfig filters the provided values down into only the values
+// required by the specified configuration.
 //
-// Crucially, it differs from buildInputVariablesForTest in that the returned
-// input values include all variables available even if they are not defined
-// within the config. This allows the assertions to refer to variables defined
-// solely within the test file, and not only those within the configuration.
-//
-// In addition, it modifies the provided config so that any variables that are
-// available are also defined in the config. It returns a function that resets
-// the config which must be called so the config can be reused going forward.
-func (runner *TestFileRunner) prepareInputVariablesForAssertions(config *configs.Config, run *moduletest.Run, file *moduletest.File, references []*addrs.Reference) (terraform.InputValues, func(), tfdiags.Diagnostics) {
-	var diags tfdiags.Diagnostics
-
-	// process is a helper function that converts an unparsed variable into an
-	// input value. All the various input formats share this logic so we extract
-	// it out here.
-	process := func(name string, value backend.UnparsedVariableValue, reference *addrs.Reference) (*terraform.InputValue, tfdiags.Diagnostics) {
-		if config, exists := config.Module.Variables[name]; exists {
-			variable, diags := value.ParseVariableValue(config.ParsingMode)
-			if diags.HasErrors() {
-				return variable, diags
-			}
-
-			given := variable.Value
-
-			// Normally, variable values would be converted during the Terraform
-			// graph processing. But, `terraform test` assertions are not
-			// executed during the graph but after. This means the variables we
-			// create for use in the assertions must be converted here.
-
-			if config.TypeDefaults != nil && !given.IsNull() {
-				given = config.TypeDefaults.Apply(given)
-			}
-
-			converted, err := convert.Convert(given, config.ConstraintType)
-			if err != nil {
-				var subject *hcl.Range
-				if reference != nil {
-					subject = reference.SourceRange.ToHCL().Ptr()
-				}
-
-				diags = diags.Append(&hcl.Diagnostic{
-					Severity: hcl.DiagError,
-					Summary:  "Invalid value for input variable",
-					Detail:   fmt.Sprintf("The given value is not suitable for var.%s declared at %s: %s.", name, config.DeclRange.String(), err),
-					Subject:  subject,
-				})
-				return variable, diags
-			}
-
-			variable.Value = converted
-			return variable, diags
-		} else {
-
-			// If the variable isn't defined in the config, then we don't know
-			// what type it is supposed to be. So we'll just parse it as HCL and
-			// we can deduce the type that way.
-
-			return value.ParseVariableValue(configs.VariableParseHCL)
-		}
-	}
-
-	// relevant keeps track of the variables that are actually referenced by
-	// this set of assertions.
-	relevant := make(map[string]*addrs.Reference)
-	for _, reference := range references {
-		addr, ok := reference.Subject.(addrs.InputVariable)
-		if !ok {
-			// We only care about variables.
+// This function is essentially the opposite of AddVariablesToConfig which
+// makes the config match the variables rather than the variables match the
+// config.
+func (runner *TestFileRunner) FilterVariablesToConfig(config *configs.Config, values terraform.InputValues) terraform.InputValues {
+	filtered := make(terraform.InputValues)
+	for name, value := range values {
+		if _, exists := config.Module.Variables[name]; !exists {
+			// Only include values that are actually required by the config.
 			continue
 		}
 
-		relevant[addr.Name] = reference
+		filtered[name] = value
 	}
+	return filtered
+}
 
-	variables := make(terraform.InputValues)
+// AddVariablesToConfig extends the provided config to ensure it has definitions
+// for all specified variables.
+//
+// This function is essentially the opposite of FilterVariablesToConfig which
+// makes the variables match the config rather than the config match the
+// variables.
+func (runner *TestFileRunner) AddVariablesToConfig(config *configs.Config, variables terraform.InputValues) func() {
 
-	// Now, we're going to process the various different sources of variables
-	// and turn them into input values that our test context can read.
-
-	// First, we'll process the global variables.
-
-	for name, value := range runner.Suite.GlobalVariables {
-		variable, variableDiags := process(name, value, relevant[name])
-		diags = diags.Append(variableDiags)
-		if variable != nil {
-			variables[name] = variable
-		}
-	}
-
-	// Second, we'll process the variables from the file.
-
-	if file != nil {
-		for name, expr := range file.Config.Variables {
-			value := unparsedTestVariableValue{
-				expr: expr,
-			}
-
-			variable, variableDiags := process(name, value, relevant[name])
-			diags = diags.Append(variableDiags)
-			if variable != nil {
-				variables[name] = variable
-			}
-		}
-	}
-
-	// Third, we'll process the variables from the run block. We pass in the
-	// variables from the global and file level into the eval context here so
-	// that users can set run variables from file and global variables.
-
-	if run != nil {
-		skipVars := false
-
-		ctx, ctxDiags := runner.ctx(run, file, variables)
-		diags = diags.Append(ctxDiags)
-		if ctxDiags.HasErrors() {
-			// Then we won't try and actually evaluate run variables but we do
-			// keep note of them.
-			skipVars = true
-		}
-
-		for name, expr := range run.Config.Variables {
-
-			if skipVars {
-
-				// Then we had a problem with the evaluation context.
-				//
-				// We'll just make a placeholder input value so we can finish
-				// evaluating everything else. We won't end up using the
-				// placeholder values as the test will fail due to the errored
-				// diags when we build the context.
-
-				variables[name] = &terraform.InputValue{
-					Value:       cty.NilVal,
-					SourceType:  terraform.ValueFromConfig,
-					SourceRange: tfdiags.SourceRangeFromHCL(expr.Range()),
-				}
-
-				continue
-			}
-
-			value := unparsedTestVariableValue{
-				expr: expr,
-				ctx:  ctx,
-			}
-
-			variable, variableDiags := process(name, value, relevant[name])
-			diags = diags.Append(variableDiags)
-			if variable != nil {
-				variables[name] = variable
-			}
-		}
-	}
-
-	// Finally, we look for any default values from the configuration for
-	// variables that we haven't assigned a value to yet.
-
-	for name, variable := range config.Module.Variables {
-		if _, exists := variables[name]; exists {
-			// Then we don't want to apply the default for this variable as we
-			// already have a value.
-			continue
-		}
-
-		if variable.Default != cty.NilVal {
-			variables[name] = &terraform.InputValue{
-				Value:       variable.Default,
-				SourceType:  terraform.ValueFromConfig,
-				SourceRange: tfdiags.SourceRangeFromHCL(variable.DeclRange),
-			}
-		}
-	}
-
-	// Now we're going to do a some modifications to the config.
-	//
 	// If we have got variable values from the test file we need to make sure
 	// they have an equivalent entry in the configuration. We're going to do
 	// that dynamically here.
@@ -1286,8 +1156,6 @@ func (runner *TestFileRunner) prepareInputVariablesForAssertions(config *configs
 		currentVars[name] = variable
 	}
 
-	// Next, let's go through our entire inputs and add any that aren't already
-	// defined into the config.
 	for name, value := range variables {
 		if _, exists := config.Module.Variables[name]; exists {
 			continue
@@ -1301,200 +1169,22 @@ func (runner *TestFileRunner) prepareInputVariablesForAssertions(config *configs
 		}
 	}
 
-	// We return our input values, a function that will reset the variables
-	// within the config so it can be used again, and any diagnostics reporting
-	// variables that we couldn't parse.
-
-	return variables, func() {
+	// We return a function that will reset the variables within the config so
+	// it can be used again.
+	return func() {
 		config.Module.Variables = currentVars
-	}, diags
+	}
 }
 
-// EvalCtx returns an hcl.EvalContext that allows the variables blocks within
-// run blocks to evaluate references to the outputs from other run blocks.
-func (runner *TestFileRunner) ctx(run *moduletest.Run, file *moduletest.File, availableVariables terraform.InputValues) (*hcl.EvalContext, tfdiags.Diagnostics) {
-	var diags tfdiags.Diagnostics
-
-	// First, let's build the set of available run blocks.
-
-	availableRunBlocks := make(map[string]*terraform.TestContext)
-	runs := make(map[string]cty.Value)
-	for _, run := range file.Runs {
-		name := run.Name
-
-		attrs := make(map[string]cty.Value)
-		if ctx, exists := runner.PriorStates[name]; exists {
-			// We have executed this run block previously, therefore it is
-			// available as a reference at this point in time.
-			availableRunBlocks[name] = ctx
-
-			for name, config := range ctx.Config.Module.Outputs {
-				output := ctx.State.OutputValue(addrs.AbsOutputValue{
-					OutputValue: addrs.OutputValue{
-						Name: name,
-					},
-					Module: addrs.RootModuleInstance,
-				})
-
-				var value cty.Value
-				switch {
-				case output == nil:
-					// This means the run block returned null for this output.
-					// It is likely this will produce an error later if it is
-					// referenced, but users can actually specify that null
-					// is an acceptable value for an input variable so we won't
-					// actually raise a fuss about this at all.
-					value = cty.NullVal(cty.DynamicPseudoType)
-				case output.Value.IsNull() || output.Value == cty.NilVal:
-					// This means the output value was returned as (known after
-					// apply). If this is referenced it always an error, we
-					// can't handle this in an appropriate way at all. For now,
-					// we just mark it as unknown and then later we check and
-					// resolve all the references. We'll raise an error at that
-					// point if the user actually attempts to reference a value
-					// that is unknown.
-					value = cty.DynamicVal
-				default:
-					value = output.Value
-				}
-
-				if config.Sensitive || (output != nil && output.Sensitive) {
-					value = value.Mark(marks.Sensitive)
-				}
-
-				attrs[name] = value
-			}
-
-			runs[name] = cty.ObjectVal(attrs)
-
-			continue
-		}
-
-		// We haven't executed this run block yet, therefore it is not available
-		// as a reference at this point in time.
-		availableRunBlocks[name] = nil
+// initVariables initialises the globalVariables within the test runner by
+// merging the global variables from the test suite into the variables from
+// the file.
+func (runner *TestFileRunner) initVariables(file *moduletest.File) {
+	runner.globalVariables = make(map[string]backend.UnparsedVariableValue)
+	for name, value := range runner.Suite.GlobalVariables {
+		runner.globalVariables[name] = value
 	}
-
-	// Second, let's build the set of available variables.
-
-	vars := make(map[string]cty.Value)
-	for name, variable := range availableVariables {
-		vars[name] = variable.Value
+	for name, expr := range file.Config.Variables {
+		runner.globalVariables[name] = unparsedTestVariableValue{expr}
 	}
-
-	// Third, let's do some basic validation over the references.
-
-	for _, value := range run.Config.Variables {
-		refs, refDiags := lang.ReferencesInExpr(addrs.ParseRefFromTestingScope, value)
-		diags = diags.Append(refDiags)
-		if refDiags.HasErrors() {
-			continue
-		}
-
-		for _, ref := range refs {
-			if addr, ok := ref.Subject.(addrs.Run); ok {
-				ctx, exists := availableRunBlocks[addr.Name]
-
-				if !exists {
-					// Then this is a made up run block.
-					diags = diags.Append(&hcl.Diagnostic{
-						Severity: hcl.DiagError,
-						Summary:  "Reference to unknown run block",
-						Detail:   fmt.Sprintf("The run block %q does not exist within this test file. You can only reference run blocks that are in the same test file and will execute before the current run block.", addr.Name),
-						Subject:  ref.SourceRange.ToHCL().Ptr(),
-					})
-
-					continue
-				}
-
-				if ctx == nil {
-					// This run block exists, but it is after the current run block.
-					diags = diags.Append(&hcl.Diagnostic{
-						Severity: hcl.DiagError,
-						Summary:  "Reference to unavailable run block",
-						Detail:   fmt.Sprintf("The run block %q is not available to the current run block. You can only reference run blocks that are in the same test file and will execute before the current run block.", addr.Name),
-						Subject:  ref.SourceRange.ToHCL().Ptr(),
-					})
-
-					continue
-				}
-
-				value, valueDiags := ref.Remaining.TraverseRel(runs[addr.Name])
-				diags = diags.Append(valueDiags)
-				if valueDiags.HasErrors() {
-					// This means the reference was invalid somehow, we've
-					// already added the errors to our diagnostics though so
-					// we'll just carry on.
-					continue
-				}
-
-				if !value.IsWhollyKnown() {
-					// This is not valid, we cannot allow users to pass unknown
-					// values into run blocks. There's just going to be
-					// difficult and confusing errors later if this happens.
-
-					if ctx.Run.Config.Command == configs.PlanTestCommand {
-						// Then the user has likely attempted to use an output
-						// that is (known after apply) due to the referenced
-						// run block only being a plan command.
-						diags = diags.Append(&hcl.Diagnostic{
-							Severity: hcl.DiagError,
-							Summary:  "Reference to unknown value",
-							Detail:   fmt.Sprintf("The value for %s is unknown. Run block %q is executing a \"plan\" operation, and the specified output value is only known after apply.", ref.DisplayString(), addr.Name),
-							Subject:  ref.SourceRange.ToHCL().Ptr(),
-						})
-
-						continue
-					}
-
-					// Otherwise, this is a bug in Terraform. We shouldn't be
-					// producing (known after apply) values during apply
-					// operations.
-					diags = diags.Append(&hcl.Diagnostic{
-						Severity: hcl.DiagError,
-						Summary:  "Reference to unknown value",
-						Detail:   fmt.Sprintf("The value for %s is unknown; This is a bug in Terraform, please report it.", ref.DisplayString()),
-						Subject:  ref.SourceRange.ToHCL().Ptr(),
-					})
-				}
-
-				continue
-			}
-
-			if addr, ok := ref.Subject.(addrs.InputVariable); ok {
-				if _, exists := vars[addr.Name]; !exists {
-					// This variable reference doesn't exist.
-					diags = diags.Append(&hcl.Diagnostic{
-						Severity: hcl.DiagError,
-						Summary:  "Reference to unavailable variable",
-						Detail:   fmt.Sprintf("The input variable %q is not available to the current run block. You can only reference variables defined at the file or global levels when populating the variables block within a run block.", addr.Name),
-						Subject:  ref.SourceRange.ToHCL().Ptr(),
-					})
-
-					continue
-				}
-
-				// Otherwise, we're good. This is an acceptable reference.
-				continue
-			}
-
-			// You can only reference run blocks and variables from the run
-			// block variables.
-			diags = diags.Append(&hcl.Diagnostic{
-				Severity: hcl.DiagError,
-				Summary:  "Invalid reference",
-				Detail:   "You can only reference earlier run blocks, file level, and global variables while defining variables from inside a run block.",
-				Subject:  ref.SourceRange.ToHCL().Ptr(),
-			})
-		}
-	}
-
-	// Finally, we can just populate our hcl.EvalContext.
-
-	return &hcl.EvalContext{
-		Variables: map[string]cty.Value{
-			"run": cty.ObjectVal(runs),
-			"var": cty.ObjectVal(vars),
-		},
-	}, diags
 }